+++ test/support/combustion.rb
@@ -8,9 +8,13 @@ if Rails.version >= "6.0"
   Lockbox.encrypts_action_text_body
 end
 
-Combustion.initialize! *components do
+Combustion.initialize!(*components) do
   if ActiveRecord::VERSION::MAJOR < 6 && config.active_record.sqlite3.respond_to?(:represent_boolean_as_integer)
     config.active_record.sqlite3.represent_boolean_as_integer = true
+  end
+
+  if ActiveRecord::VERSION::MAJOR >= 7
+    config.active_record.legacy_connection_handling = false
   end
 
   config.logger = $logger
@@ -21,5 +25,9 @@ Combustion.initialize! *components do
 
   if defined?(ActiveStorage)
     config.active_storage.service = :test
+
+    if ActiveRecord::VERSION::MAJOR >= 7
+      config.active_storage.replace_on_assign_to_many = true
+    end
   end
 end